boot.means	R Documentation

Example dataframe of mean Fst and heterozygosity from the bootstrapping for the bins.

Description

Example dataframe of mean Fst and Ht from bootstrapping. This file contains a dataframe with 22 columns and 5 rows. The columns are mean Ht, mean Fst, the number of loci in the bin, and the lower and upper bounds for each bin.

Usage

boot.means

Format

data.frame

Source

boot.means<-fst.boot.means(boot.out)

References

See Flanagan & Jones