boot.out	R Documentation

Example bootstrap output from numerical simulations

Description

Example bootstrap output from numerical simulations It was generated by using a numerical analysis with Nm = 10, 75 demes, and 5 population samples taken. No selection was imposed. Ten bootstrap replicates were run on the example.genepop dataframe. This is a data.frame of lists. The first list is Fsts, which is a list of dataframes, each dataframe containing two columns: Ht and Fst. The second list is a list of data.frames, each containing the bins used in the bootstrapping module. The final list is a list of list. Each of the lists is a list of the upper and lower confidence intervals.

Usage

boot.out

Format

data.frame

Source

From bootstrapping 10 reps over the dataframe gpop.

References

See Flanagan & Jones