boot.out.list	R Documentation

Example list of CI matrices from bootstrap output from numerical simulations

Description

Example list of CI matrices from bootstrap output from numerical simulations The data were generated using a numerical analysis with Nm = 10, 75 demes, and 5 population samples taken. No selection was imposed. Ten bootstrap replicates were run on gpop. This is a lists of matrices containing the 10 sets of values from the 95 percent confidence intervals.

Usage

boot.out.list

Format

list

Source

From bootstrapping 10 reps over the dataframe gpop.

References

See Flanagan & Jones