calc.actual.fst	R Documentation

This calcualtes global Fsts from a genepop dataframe.

Description

This calcualtes global Fsts from a genepop dataframe. This does not include bootstrapping.

Usage

calc.actual.fst(df, fst.choice="N")

Arguments

df
Provide the genepop dataframe (from my.read.genepop).

fst.choice
Specify which type of fst calculation should be used. See fst.options.print for the choices.

Value

fsts
This returns a dataframe with Locus, Ht, and Fst characters.