calc.fst	R Documentation

This calculates Fst.

Description

This calculates Fst. The caluclation is done as (Ht-Hs)/Ht, where Ht is the expected heterozygosity for all populations and Hs is the expected heterozygosity for each population. This calculation is used in bootstrapping functions.

Usage

calc.fst(df, i)

Arguments

df
A dataframe containing the genepop information, where the first column is the population ID.

i
Column number containing genotype information.

Value

ht
The expected heterozygosity under Hardy-Weinberg expectations. This is a single numerical value.

fst
The calculated Fst value for this locus.