ci.means	R Documentation

This calculates the average confidence intervals from multiple bootstrap outputs.

Description

This calculates the mean upper and lower confidence intervals from a list of bootstrap CI matrices.

Usage

ci.means(boot.out.list)

Arguments

boot.out.list
A list of matrices. Each matrix is the CIs from fst.boot (boot.out[[3]]).

Value

avg.cil
A list of the average lower CI values

avg.ciu
A list of the average upper CI values