find.outliers	R Documentation

This identifies all of the SNPs outside of the confidence intervals in the dataset.

Description

This identifies all of the SNPs outside of the confidence intervals in the dataset.

Usage

find.outliers(df, ci.df = NULL, boot.out = NULL, file.name = NULL)

Arguments

df
Provide the dataframe with Ht and Fst values.

boot.out
Bootstrap output. You must either provide this or a list of confidence interval values.

ci.df
List of confidence intervals. You must either provide this or bootstrap output.

file.name
You may provide a file name to output the outliers to a csv file. Otherwise, the function will only return the 95 and 99 percent outliers.

Value

out95
A list of the 95 percent outlier loci

out99
A list of the 99 percent outlier loci