fst.boot.means	R Documentation

Calculates mean values within the bins.

Description

This calculates mean heterozygosity and Fst values for each bin used in bootstrapping.

Usage

fst.boot.means(boot.out[[1]])

Arguments

boot.out.means
The first item in the output lists from fst.boot.

Value

bmu
A dataframe containing four columns: heterozygosity Fst the number of loci in the bin the lower Ht value for the bin and the upper Ht value for the bin.