fst.boot.onecol	R Documentation

This bootstraps across all individuals to calculate a bootstrapped Fst for a randomly-sampled locus.

Description

This calculates Fst using calc.fst. It randomly selects a column containing genotype information for all individuals. It then calculates Fst and Ht for that locus.

Usage

fst.boot.onecol(df)

Arguments

df
A dataframe containing the genepop information, where the first column is the population ID.

Value

ht.fst
A vector containin Ht and Fst