fsts.wc	R Documentation

Example fst calculations from a genepop file.

Description

Example fst calculations from a genepop file. The Fsts were calculated using the Weir and Cockerham (1993) calculation. The original data were generated by using a numerical analysis with Nm = 10, 75 demes, and 5 population samples taken. No selection was imposed. The fsts were calculated using calc.actual.fst(gpop) This file contains a dataframe with 2000 columns and 3 rows. The first column is the Locus ID, the second column is the Ht for that locus, and the third column is the Fst for that locus.

Usage

fsts.wc

Format

data.frame

Source

Generated by numerical analysis

References

See Flanagan & Jones