p.boot	R Documentation

Calculates mean values within the bins.

Description

This calculates mean heterozygosity and Fst values for each bin used in bootstrapping.

Usage

p.boot(actual.fsts, boot.out)

Arguments

actual.fsts
The first item in the output lists from fst.boot.

boot.out
The output from a bootstrapping run. Either supply this or boot.means.

boot.means
The output from fst.boot.means. Either supply this or bootstrapping output.

Value

pvals
A numeric containing uncorrected p-values for each locus. The names attribute are the locus names.