calc.allele.freq	R Documentation

This calculates allele frequencies.

Description

This calculates allele frequencies from a list of genotypes.

Usage

calc.allele.freq(genotypes)

Arguments

genotypes
A list of genotypes.

Value

obs.af
A list of observed allele frequencies in the genotypes list.