find.outliers	R Documentation

This identifies all of the SNPs outside of the confidence intervals in the dataset.

Description

This identifies all of the SNPs outside of the confidence intervals in the dataset.

Usage

find.outliers(df, boot.out, ci.df = NULL, file.name = NULL)

Arguments

df
Provide the dataframe with Ht and Fst values.

boot.out
Bootstrap output. You must provide this.

ci.df
List of confidence intervals. You may provide this in addition to bootstrap output to save a small amount of time.

file.name
You may provide a file name to output the outliers to a csv file. Otherwise, the function will only return the outliers.

Value

out
A list of the outlier loci